jl_to_psql/requirements.txt
numpy==1.19.5
pandas==1.1.5
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0

